requirements.txt
openai
langchain
fastapi
uvicorn
chromadb
faiss-cpu
jinja2
python-dotenv
pyyaml
requests
